chore: move logger into `/internal`
