+++ internal/logger/logger.go
@@ -1,4 +1,4 @@
-package utils
+package logger
 
 import (
 	"go.elastic.co/ecszap"
